Integrated passport into profile.js
