+++ client/src/pages/Profile.js
@@ -1,12 +1,88 @@
-import React from 'react';
+import React, { useState } from 'react';
 import Header from '../components/Header';
+// import './App.css';
+import Axios from 'axios';
+
+// function Profile() {
+//     return(
+//         <div className="container container-fluid">
+//             <Header headerName={ "Create a Profile" }/>
+//         </div>
+//     )
+// }
 
 function Profile() {
-    return(
-        <div className="container container-fluid">
-            <Header headerName={ "Create a Profile" }/>
+    const [registerUsername, setRegisterUsername] = useState('');
+    const [registerPassword, setRegisterPassword] = useState('');
+    const [loginUsername, setLoginUsername] = useState('');
+    const [loginPassword, setLoginPassword] = useState('');
+    const [data, setData] = useState(null);
+    const register = () => {
+      Axios({
+        method: 'POST',
+        data: {
+          username: registerUsername,
+          password: registerPassword,
+        },
+        withCredentials: true,
+        url: 'http://localhost:3000/register',
+      }).then((res) => console.log(res));
+    };
+    const login = () => {
+      Axios({
+        method: 'POST',
+        data: {
+          username: loginUsername,
+          password: loginPassword,
+        },
+        withCredentials: true,
+        url: 'http://localhost:3000/login',
+      }).then((res) => console.log(res));
+    };
+    const getUser = () => {
+      Axios({
+        method: 'GET',
+        withCredentials: true,
+        url: 'http://localhost:3000/user',
+      }).then((res) => {
+        setData(res.data);
+        console.log(res.data);
+      });
+    };
+    return (
+      <div className='App'>
+        <div>
+        <Header headerName={ "Create a Profile" }/>
+          <input
+            placeholder='username'
+            onChange={(e) => setRegisterUsername(e.target.value)}
+          />
+          <input
+            placeholder='password'
+            onChange={(e) => setRegisterPassword(e.target.value)}
+          />
+          <button onClick={register}>Submit</button>
         </div>
-    )
-}
-
+  
+        <div>
+        <Header headerName={ "Login" }/>
+          <input
+            placeholder='username'
+            onChange={(e) => setLoginUsername(e.target.value)}
+          />
+          <input
+            placeholder='password'
+            onChange={(e) => setLoginPassword(e.target.value)}
+          />
+          <button onClick={login}>Submit</button>
+        </div>
+  
+        <div>
+        <Header headerName={ "Welcome Back!" }/>
+          <button onClick={getUser}>Submit</button>
+          {data ? <h1>Welcome Back {data.username}</h1> : null}
+        </div>
+      </div>
+    );
+  }
 export default Profile;
